--- DB/banatec.log
@@ -1,31 +1,0 @@
-/*C1*/SET SCHEMA SYSTEM_LOBS
-INSERT INTO BLOCKS VALUES(0,2147483647,0)
-COMMIT
-/*C3*/SET SCHEMA PUBLIC
-CREATE TABLE asignacion (\u000a  id_asig INT NOT NULL,\u000a  costo INT NOT NULL,\u000a  trabajador INT NOT NULL,\u000a  asignador INT NOT NULL,\u000a  lote INT NOT NULL,\u000a  labor INT NOT NULL,\u000a  f_inicio DATE NOT NULL,\u000a  f_fin DATE NOT NULL,\u000a  PRIMARY KEY (id_asig)\u000a)
-CREATE TABLE trabajadores (\u000a  cedula_tra INT NOT NULL,\u000a  nombre_t VARCHAR(25) NOT NULL,\u000a  apellido_t VARCHAR(25) NOT NULL,\u000a  telefono_t VARCHAR(12) DEFAULT NULL,\u000a  edad_t VARCHAR(2) DEFAULT NULL,\u000a  direccion_t VARCHAR(50) DEFAULT NULL,\u000a  PRIMARY KEY (cedula_tra)\u000a)
-CREATE TABLE administradores (\u000a  cedula_admin INT NOT NULL,\u000a  nombre_a VARCHAR(25) NOT NULL,\u000a  apellido_a VARCHAR(25) NOT NULL,\u000a  correo_a VARCHAR(40) NOT NULL,\u000a  telefono_a VARCHAR(12) DEFAULT NULL,\u000a  edad_a VARCHAR(2) DEFAULT NULL,\u000a  direccion_a VARCHAR(50) DEFAULT NULL,\u000a  contrase\u00f1a VARCHAR(50) NOT NULL,\u000a  PRIMARY KEY (cedula_admin)\u000a)
-CREATE TABLE labores (\u000a  id_labor INT NOT NULL,\u000a  nombre_l VARCHAR(40) DEFAULT NULL,\u000a  descripcion_l VARCHAR(60) DEFAULT NULL,\u000a  PRIMARY KEY (id_labor)\u000a)
-CREATE TABLE cintaje (\u000a  id_encintaje INT NOT NULL,\u000a  lote INT DEFAULT NULL,\u000a  amarillo INT DEFAULT NULL,\u000a  naranja INT DEFAULT NULL,\u000a  rojo INT DEFAULT NULL,\u000a  blanco INT DEFAULT NULL,\u000a  negro INT DEFAULT NULL,\u000a  azul INT DEFAULT NULL,\u000a  morado INT DEFAULT NULL,\u000a  cafe INT DEFAULT NULL,\u000a  total INT DEFAULT NULL,\u000a  PRIMARY KEY (id_encintaje)\u000a)
-CREATE TABLE cosechas (\u000a  id_cosecha INT NOT NULL ,\u000a  lote INT DEFAULT NULL,\u000a  cantidad INT DEFAULT NULL,\u000a  fecha DATE DEFAULT NULL,\u000a  PRIMARY KEY (id_cosecha) \u000a)
-CREATE TABLE lotes (\u000a  id_lote INT NOT NULL,\u000a  descripcion VARCHAR(30) DEFAULT NULL,\u000a  PRIMARY KEY (id_lote)\u000a)
-ALTER TABLE asignacion ADD FOREIGN KEY(trabajador) REFERENCES trabajadores(cedula_tra)
-ALTER TABLE asignacion ADD FOREIGN KEY(asignador) REFERENCES administradores(cedula_admin)
-ALTER TABLE asignacion ADD FOREIGN KEY(labor) REFERENCES labores(id_labor)
-ALTER TABLE asignacion ADD FOREIGN KEY(lote) REFERENCES lotes(id_lote)
-ALTER TABLE cintaje ADD FOREIGN KEY(lote) REFERENCES lotes(id_lote)
-ALTER TABLE cosechas ADD FOREIGN KEY(lote) REFERENCES lotes(id_lote)
-INSERT INTO ADMINISTRADORES VALUES(111,'aaa','aaa','a.com','2323','32','2332e','123')
-INSERT INTO ADMINISTRADORES VALUES(123,'nestor','ortiz','n.com','123','23','robledo','123')
-INSERT INTO ADMINISTRADORES VALUES(1332234,'brayan','marin','brayan123@gmail.com','234454','20','cra 34 45-56','123456')
-INSERT INTO ADMINISTRADORES VALUES(102993922,'James','Arboleda','james@gmail.com','231232','21','Medellin','12345')
-INSERT INTO ADMINISTRADORES VALUES(1216723664,'nestor','ortiz','nestorortizcardona@gmail.com','3126306222','23','robledo','123456')
-COMMIT
-INSERT INTO LABORES VALUES(3,'bolsaje','El trabajador debe embolsar las matas')
-INSERT INTO LABORES VALUES(4,'encintar','El trabajador poner cinta de color a la mata')
-COMMIT
-INSERT INTO LOTES VALUES(1,'lote 1')
-COMMIT
-INSERT INTO TRABAJADORES VALUES(2344,'james','arboleda','4678434','20','cra 56 \u00b754-66')
-INSERT INTO TRABAJADORES VALUES(71834567,'sandra','quintero','3214569345','36','avenida las palmas')
-COMMIT
