--- DB/banatec.properties
@@ -1,5 +1,0 @@
-#HSQL Database Engine 2.5.1
-#Mon Sep 14 18:43:49 COT 2020
-version=2.5.1
-modified=yes
-tx_timestamp=1
