+++ nbproject/project.properties
@@ -35,13 +35,19 @@ endorsed.classpath=
 excludes=
 file.reference.hsqldb.jar=lib\\hsqldb.jar
 file.reference.jcalendar-1.1.4.jar=lib/jcalendar-1.1.4.jar
+file.reference.jcommon-1.0.17.jar=lib\\jcommon-1.0.17.jar
+file.reference.jfreechart-1.0.14.jar=lib\\jfreechart-1.0.14.jar
+file.reference.sqlite-jdbc-3.23.1.jar=lib\\sqlite-jdbc-3.23.1.jar
 includes=**
 jar.compress=false
 javac.classpath=\
     ${libs.MySQLDriver.classpath}:\
     ${libs.absolutelayout.classpath}:\
     ${file.reference.jcalendar-1.1.4.jar}:\
-    ${file.reference.hsqldb.jar}
+    ${file.reference.hsqldb.jar}:\
+    ${file.reference.jcommon-1.0.17.jar}:\
+    ${file.reference.jfreechart-1.0.14.jar}:\
+    ${file.reference.sqlite-jdbc-3.23.1.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
