--- test/vista/AdministradoresTest.java
@@ -1,112 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package vista;
-
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-/**
- *
- * @author SHAGGY
- */
-public class AdministradoresTest {
-    
-    public AdministradoresTest() {
-    }
-    
-    @BeforeClass
-    public static void setUpClass() {
-    }
-    
-    @AfterClass
-    public static void tearDownClass() {
-    }
-    
-    @Before
-    public void setUp() {
-    }
-    
-    @After
-    public void tearDown() {
-    }
-
-    /**
-     * Test of lista method, of class Administradores.
-     */
-    @Test
-    public void testLista() {
-        System.out.println("lista");
-        Administradores instance = new Administradores();
-        instance.lista();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of agregar method, of class Administradores.
-     */
-    @Test
-    public void testAgregar() {
-        System.out.println("agregar");
-        Administradores instance = new Administradores();
-        instance.agregar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of actualizar method, of class Administradores.
-     */
-    @Test
-    public void testActualizar() {
-        System.out.println("actualizar");
-        Administradores instance = new Administradores();
-        instance.actualizar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of eliminar method, of class Administradores.
-     */
-    @Test
-    public void testEliminar() {
-        System.out.println("eliminar");
-        Administradores instance = new Administradores();
-        instance.eliminar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of cancelar method, of class Administradores.
-     */
-    @Test
-    public void testCancelar() {
-        System.out.println("cancelar");
-        Administradores instance = new Administradores();
-        instance.cancelar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of limpia method, of class Administradores.
-     */
-    @Test
-    public void testLimpia() {
-        System.out.println("limpia");
-        Administradores instance = new Administradores();
-        instance.limpia();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-    
-}
