--- test/vista/CintajeTest.java
@@ -1,138 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package vista;
-
-import javax.swing.JComboBox;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-/**
- *
- * @author SHAGGY
- */
-public class CintajeTest {
-    
-    public CintajeTest() {
-    }
-    
-    @BeforeClass
-    public static void setUpClass() {
-    }
-    
-    @AfterClass
-    public static void tearDownClass() {
-    }
-    
-    @Before
-    public void setUp() {
-    }
-    
-    @After
-    public void tearDown() {
-    }
-
-    /**
-     * Test of lista method, of class Cintaje.
-     */
-    @Test
-    public void testLista() {
-        System.out.println("lista");
-        Cintaje instance = new Cintaje();
-        instance.lista();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of agregar method, of class Cintaje.
-     */
-    @Test
-    public void testAgregar() {
-        System.out.println("agregar");
-        Cintaje instance = new Cintaje();
-        instance.agregar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of sumar method, of class Cintaje.
-     */
-    @Test
-    public void testSumar() {
-        System.out.println("sumar");
-        Cintaje instance = new Cintaje();
-        instance.sumar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of actualizar method, of class Cintaje.
-     */
-    @Test
-    public void testActualizar() {
-        System.out.println("actualizar");
-        Cintaje instance = new Cintaje();
-        instance.actualizar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of eliminar method, of class Cintaje.
-     */
-    @Test
-    public void testEliminar() {
-        System.out.println("eliminar");
-        Cintaje instance = new Cintaje();
-        instance.eliminar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of limpiar method, of class Cintaje.
-     */
-    @Test
-    public void testLimpiar() {
-        System.out.println("limpiar");
-        Cintaje instance = new Cintaje();
-        instance.limpiar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of cancelar method, of class Cintaje.
-     */
-    @Test
-    public void testCancelar() {
-        System.out.println("cancelar");
-        Cintaje instance = new Cintaje();
-        instance.cancelar();
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-    /**
-     * Test of cargar_lotes method, of class Cintaje.
-     */
-    @Test
-    public void testCargar_lotes() {
-        System.out.println("cargar_lotes");
-        JComboBox cbox_lotes = null;
-        Cintaje instance = new Cintaje();
-        instance.cargar_lotes(cbox_lotes);
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-    
-}
